[change components structure]
