--- frontend-next/templates/components/footer/styles.js
@@ -1,122 +1,0 @@
-exports.default = {
-  followBlock: {
-    minWidth: '100%',
-    borderCollapse: 'collapse',
-    msoTableLspace: '0pt',
-    msoTableRspace: '0pt',
-    msTextSizeAdjust: '100%',
-    WebkitTextSizeAdjust: '100%',
-  },
-  followBlockInner: {
-    padding: '9px',
-    msoLineHeightRule: 'exactly',
-    msTextSizeAdjust: '100%',
-    WebkitTextSizeAdjust: '100%',
-  },
-  followContentContainer: {
-    minWidth: '100%',
-    borderCollapse: 'collapse',
-    msoTableLspace: '0pt',
-    msoTableRspace: '0pt',
-    msTextSizeAdjust: '100%',
-    WebkitTextSizeAdjust: '100%',
-  },
-  tdCenter: {
-    paddingLeft: '9px',
-    paddingRight: '9px',
-    msoLineHeightRule: 'exactly',
-    msTextSizeAdjust: '100%',
-    WebkitTextSizeAdjust: '100%',
-  },
-  followContent: {
-    minWidth: '100%',
-    borderCollapse: 'collapse',
-    msoTableLspace: '0pt',
-    msoTableRspace: '0pt',
-    msTextSizeAdjust: '100%',
-    WebkitTextSizeAdjust: '100%',
-  },
-  tdCenterTop: {
-    paddingTop: '9px',
-    paddingRight: '9px',
-    paddingLeft: '9px',
-    msoLineHeightRule: 'exactly',
-    msTextSizeAdjust: '100%',
-    WebkitTextSizeAdjust: '100%',
-  },
-  tableCenter: {
-    borderCollapse: 'collapse',
-    msoTableLspace: '0pt',
-    msoTableRspace: '0pt',
-    msTextSizeAdjust: '100%',
-    WebkitTextSizeAdjust: '100%',
-  },
-  tdCenterTop2: {
-    msoLineHeightRule: 'exactly',
-    msTextSizeAdjust: '100%',
-    WebkitTextSizeAdjust: '100%',
-  },
-  tableLeft: {
-    display: 'inline',
-    borderCollapse: 'collapse',
-    msoTableLspace: '0pt',
-    msoTableRspace: '0pt',
-    msTextSizeAdjust: '100%',
-    WebkitTextSizeAdjust: '100%',
-  },
-  followContentItemContainer: {
-    paddingRight: '10px',
-    paddingBottom: '9px',
-    msoLineHeightRule: 'exactly',
-    msTextSizeAdjust: '100%',
-    WebkitTextSizeAdjust: '100%',
-  },
-  followContentItem: {
-    borderCollapse: 'collapse',
-    msoTableLspace: '0pt',
-    msoTableRspace: '0pt',
-    msTextSizeAdjust: '100%',
-    WebkitTextSizeAdjust: '100%',
-  },
-  tdLeftMiddle: {
-    paddingTop: '5px',
-    paddingRight: '10px',
-    paddingBottom: '5px',
-    paddingLeft: '9px',
-    msoLineHeightRule: 'exactly',
-    msTextSizeAdjust: '100%',
-    WebkitTextSizeAdjust: '100%',
-  },
-  tableLeft2: {
-    borderCollapse: 'collapse',
-    msoTableLspace: '0pt',
-    msoTableRspace: '0pt',
-    msTextSizeAdjust: '100%',
-    WebkitTextSizeAdjust: '100%',
-  },
-  followIconContent: {
-    msoLineHeightRule: 'exactly',
-    msTextSizeAdjust: '100%',
-    WebkitTextSizeAdjust: '100%',
-  },
-  followSocial: {
-    msoLineHeightRule: 'exactly',
-    msTextSizeAdjust: '100%',
-    WebkitTextSizeAdjust: '100%',
-  },
-  followSocialImg: {
-    display: 'block', border: 0, height: 'auto', outline: 'none', textDecoration: 'none', msInterpolationMode: 'bicubic',
-  },
-  textBlock: {
-    minWidth: '100%', borderCollapse: 'collapse', msoTableLspace: '0pt', msoTableRspace: '0pt', msTextSizeAdjust: '100%', WebkitTextSizeAdjust: '100%',
-  },
-  textBlockInner: {
-    paddingTop: '9px', msoLineHeightRule: 'exactly', msTextSizeAdjust: '100%', WebkitTextSizeAdjust: '100%',
-  },
-  textContentContainer: {
-    maxWidth: '100%', minWidth: '100%', borderCollapse: 'collapse', msoTableLspace: '0pt', msoTableRspace: '0pt', msTextSizeAdjust: '100%', WebkitTextSizeAdjust: '100%',
-  },
-  textContent: {
-    paddingTop: 0, paddingRight: '18px', paddingBottom: '9px', paddingLeft: '18px', msoLineHeightRule: 'exactly', msTextSizeAdjust: '100%', WebkitTextSizeAdjust: '100%', wordBreak: 'break-word', color: '#656565', fontFamily: '"Source Sans Pro","Helvetica Neue", Helvetica, Arial, sans-serif', fontSize: '12px', lineHeight: '150%', textAlign: 'center',
-  },
-};
