+++ frontend-next/templates/components/footer/tableLeft/styles.js
@@ -1,0 +1,39 @@
+export const styles = {
+    tableLeft: {
+        display: 'inline',
+        borderCollapse: 'collapse',
+        msoTableLspace: '0pt',
+        msoTableRspace: '0pt',
+        msTextSizeAdjust: '100%',
+        WebkitTextSizeAdjust: '100%',
+      },
+      followContentItemContainer: {
+        paddingRight: '10px',
+        paddingBottom: '9px',
+        msoLineHeightRule: 'exactly',
+        msTextSizeAdjust: '100%',
+        WebkitTextSizeAdjust: '100%',
+      },
+      followContentItem: {
+        borderCollapse: 'collapse',
+        msoTableLspace: '0pt',
+        msoTableRspace: '0pt',
+        msTextSizeAdjust: '100%',
+        WebkitTextSizeAdjust: '100%',
+      },
+      tableLeft2: {
+        borderCollapse: 'collapse',
+        msoTableLspace: '0pt',
+        msoTableRspace: '0pt',
+        msTextSizeAdjust: '100%',
+        WebkitTextSizeAdjust: '100%',
+      },
+      followIconContent: {
+        msoLineHeightRule: 'exactly',
+        msTextSizeAdjust: '100%',
+        WebkitTextSizeAdjust: '100%',
+      },
+      followSocialImg: {
+        display: 'block', border: 0, height: 'auto', outline: 'none', textDecoration: 'none', msInterpolationMode: 'bicubic',
+      },
+}
