+++ frontend-next/templates/first/Head/Head.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Head from "next/head";
+
+const Header = () => {
+    return (
+        <div>
+            
+        </div>
+    )
+}
+
+export default Header;
