+++ frontend-next/templates/first/index.js
@@ -1,0 +1,3 @@
+import First from './First';
+
+export default First;
