+++ frontend-next/templates/first/styles.js
@@ -1,0 +1,20 @@
+export const styles = { 
+    bodyTable: {
+        borderCollapse: 'collapse', msoTableLspace: '0pt', msoTableRspace: '0pt', msTextSizeAdjust: '100%', WebkitTextSizeAdjust: '100%', height: '100%', margin: 0, padding: 0, width: '100%', backgroundColor: '#ffffff',
+      },
+      bodyCell: {
+        msoLineHeightRule: 'exactly', msTextSizeAdjust: '100%', WebkitTextSizeAdjust: '100%', height: '100%', margin: 0, padding: '10px', width: '100%', borderTop: 0,
+      },
+      templateContainer: {
+        borderCollapse: 'collapse', msoTableLspace: '0pt', msoTableRspace: '0pt', msTextSizeAdjust: '100%', WebkitTextSizeAdjust: '100%', border: 0, maxWidth: '600px !important',
+      },
+      templatePreheader: {
+        background: '#ffffff none no-repeat center/cover', msoLineHeightRule: 'exactly', msTextSizeAdjust: '100%', WebkitTextSizeAdjust: '100%', backgroundColor: '#ffffff', backgroundImage: 'none', backgroundRepeat: 'no-repeat', backgroundPosition: 'center', backgroundSize: 'cover', borderTop: 0, borderBottom: 0, paddingTop: '9px', paddingBottom: '9px',
+      },
+      templateHeader: {
+        background: '#ffffff none no-repeat center/cover', msoLineHeightRule: 'exactly', msTextSizeAdjust: '100%', WebkitTextSizeAdjust: '100%', backgroundColor: '#ffffff', backgroundImage: 'none', backgroundRepeat: 'no-repeat', backgroundPosition: 'center', backgroundSize: 'cover', borderTop: 0, borderBottom: 0, paddingTop: '9px', paddingBottom: 0,
+      },
+      templateBody: {
+        background: '#ffffff none no-repeat center/cover', msoLineHeightRule: 'exactly', msTextSizeAdjust: '100%', WebkitTextSizeAdjust: '100%', backgroundColor: '#ffffff', backgroundImage: 'none', backgroundRepeat: 'no-repeat', backgroundPosition: 'center', backgroundSize: 'cover', borderTop: 0, borderBottom: '4px solid #00ff00', paddingTop: 0, paddingBottom: '9px',
+      }
+}
